+++ react-container/src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './AuthContainer';
 export * from './ConditionalWrapper';
 export * from './LoopDetector';
+export * from './Sentry';
 export * from './TranslationWrapper';