+++ react-sentry/src/index.ts
@@ -1,1 +1,2 @@
 export * from './Sentry';
+export { SentryProps } from './utils';