+++ react-tenant-selector/src/utils/types.ts
@@ -11,7 +11,6 @@ export type SidecarConfig = {
   directoryTenantId?: string;
   disableTranslations?: boolean;
   disableLegacyLogin?: boolean;
-  disableLoopDetector?: boolean;
   helpLink?: string;
   locizeProjectId?: string;
 };